pyproject.toml
[project]
name = "scraper-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "2captcha-python>=1.5.1",
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "beautifulsoup4>=4.13.4",
    "celery[redis]>=5.5.3",
    "cloudscraper>=1.2.71",
    "dateparser>=1.2.2",
    "emoji>=2.14.1",
    "fake-useragent>=2.2.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "lxml-html-clean>=0.4.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "passlib[bcrypt]>=1.7.4",
    "playwright>=1.54.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests-html>=0.10.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
